Safari-609.1.11.1/LayoutTests/svg/animations/animation-leak-list-property-instances-expected.txt

This test checks that adding an animation to a SVG element does not leak the whole SVGDocument.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 0 is 0
